requirements-dev.txt
pydantic>=2
pytest
pytest-cov
bandit
mypy
hypothesis
email-validator>=2.1.0.post1
pyyaml>=6.0
